automation/requirements.txt
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
paramiko>=3.3.1
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
